This is my Final Assignment for the HTML & CSS
This Assignment was done for the HTML & CSS class at Mohawk College. All rights are given to the HTML & CSS class at MC, and all the images, and mp3 were given from the proffesor of the course and not my own. I wrote the code, but did not create the images, and mp3.